LayoutTests/accessibility/aria-modal-multiple-dialogs-expected.txt

This tests that aria-modal works correctly on multiple dialogs

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS backgroundAccessible() is true
PASS backgroundAccessible() is false
PASS dialog1Accessible() is true
PASS backgroundAccessible() is false
PASS dialog1Accessible() is false
PASS closeBtn.isIgnored is false
PASS backgroundAccessible() is false
PASS dialog1Accessible() is true
PASS backgroundAccessible() is true
PASS successfullyParsed is true

TEST COMPLETE
Other page content with a dummy focusable element

Display a dialog
